Add method to retrieve application based on client_id
A new method, get_application, has been created. It takes the request's validated data and retrieves an Application object based on the provided client_id. If a client_id is not provided or is invalid, it returns a Response object with an appropriate error message and status code. This function is then used in the post method to further modify the request.