Improve security by handling `client_secret` differently
The `client_secret` is now no longer accepted as part of the request data in the `RevokeTokenSerializer` class, increasing overall security. A warning is also logged if `client_secret` is detected in the request data. To compensate for this change, the `client_secret` is now retrieved from the application itself in the `RevokeTokenSerializer` view post method.