Client secret removal (#234)
* Remove client_secret field from serializers

The client_secret field has been removed from the serializers in the drf-social-oauth2 package. This change is made to enhance the security of our system, as the client_secret usually kept confidential and should not be serialized.

* Add method to retrieve application based on client_id

A new method, get_application, has been created. It takes the request's validated data and retrieves an Application object based on the provided client_id. If a client_id is not provided or is invalid, it returns a Response object with an appropriate error message and status code. This function is then used in the post method to further modify the request.

* Add logging to warn against including client_secret

A logger was set up in drf_social_oauth2/views.py to produce a warning whenever a 'client_secret' is detected in the validated request data. This measure was taken to alert about potential security risks associated with unnecessarily exposing client secrets.

* Check if client_secret is present in request.data.

* Improve security by handling `client_secret` differently

The `client_secret` is now no longer accepted as part of the request data in the `RevokeTokenSerializer` class, increasing overall security. A warning is also logged if `client_secret` is detected in the request data. To compensate for this change, the `client_secret` is now retrieved from the application itself in the `RevokeTokenSerializer` view post method.

* Add verbosity to pytest in docker-compose tests

* Update tests and fixtures in drf-social-oauth2

Updated the drf-social-oauth2 test fixtures to include client_id in the app and modified the save function to get the first Application object for a user. The added client_id enhances test coverage and getting the first Application object handles scenarios where a User may associate with more than one Application.

* Remove token field from RevokeTokenSerializer

The token field, which previously had a character limit of 5000, was removed from the RevokeTokenSerializer class in drf_social_oauth2/serializers.py. This simplifies the serializer as this field is no longer needed.

* Update test cases in OAuth2 endpoints and views

This commit updates the test cases in OAuth2 endpoints to use the first Application object associated with a user rather than getting it directly. It also extends tests for Application retrieval in the views to cover various scenarios including correct, incorrect, and missing client_id. Changes in test data have also been reflected, streamlining correct client_ids. Some redundant token tests in views have been removed.

* Extract token from authorization header

Our application now extracts the token from the authorization header in the RevokeToken endpoint. This change has been made to facilitate a more secure token revoke in the drf-social-oauth2 authentication, where the token in use is pulled directly from the header instead of being passed through request data.

* Update user application retrieval in test fixtures

This change updates the method of retrieving a user's application in the test fixtures of the drf-social-oauth2 project.

* Update OAuth2 tests and rename get_application function

Removed 'client_secret' from OAuth2 tests to simplify token requests. Updated 'get_application' function to 'get_application_or_400' across the test_views.py file. Also, added a test case for token revocation upon receiving an invalid token endpoint for enhanced security checks.

* Rename get_application to get_application_or_400

The method get_application has been renamed to get_application_or_400 to more accurately reflect its behavior. This change reflects across the entire module, impacting both the ConvertTokenSerializer and RevokeTokenSerializer methods, which now call the renamed method. The new name indicates the method returns an HTTP 400 error if the application is not retrieved properly.

* Update database settings in test environment

The hardcoded database parameters in the test_settings.py file have been replaced with environmental variables. This improves security and enables easier configuration in various deployment environments.

* Refactor get_application to return None instead of 400

The get_application method has been refactored to return None when the application doesn't exist or the client_id is missing. The check to determine if the application is not found has been moved to ConvertTokenSerializer and RevokeTokenSerializer methods. Corresponding test cases have also been updated to reflect the changes.

* Update version to 3.0.0 in various files

The version number of drf-social-oauth2 has been updated to 3.0.0 from 2.3.0 in the __init__.py, the conf.py in the docs, and the installation instructions in installation.rst. This change reflects the new additional features and improvements in the new version, which are yet to be documented.

* Add new application image.

* Add LinkedIn setup instructions to documentation.

* Remove 'client_secret' from API definitions

The 'client_secret' field has been removed from the API definitions in the 'api.yaml' file. This change affects both the token request and revoke token request definitions. The removal of this field simplifies the API usage and increases security by no longer requiring the client secret for these operations.