Update package version to 3.0.1 (#236)
* Update package version to 3.0.1

The version number in three places: __init__.py, conf.py in the docs source, and the installation instructions in the docs source has been updated from 3.0.0 to 3.0.1 to reflect the latest changes.

* Update database settings and test view

Detailed test settings for the database have been added, including default credentials and database name. Test local databases are now defaulting to PORT 666 if no environment variable is set. Additionally, the id used for forcing authentication in the test view has been changed and as a result, the expected response status code is also updated.